refactor: Consolidate `Summary` in BuildsCore
This removes the unnecessary `Summary` extension in the Builds target.
